move some utils from getLabware.js into helpers/
