+++ shared-data/js/helpers/index.js
@@ -1,12 +1,16 @@
 // @flow
 import assert from 'assert'
-import { getLabwareDefURI } from '../getLabware'
+import uniq from 'lodash/uniq'
 import type { LabwareDefinition2 } from '../types'
+
 export { canPipetteUseLabware } from './canPipetteUseLabware'
 export { getWellNamePerMultiTip } from './getWellNamePerMultiTip'
 export { default as getWellTotalVolume } from './getWellTotalVolume'
 export { default as wellIsRect } from './wellIsRect'
 export * from './volume'
+
+export const getLabwareDefURI = (def: LabwareDefinition2): string =>
+  `${def.namespace}/${def.parameters.loadName}/${def.version}`
 
 export const getLabwareDisplayName = (labwareDef: LabwareDefinition2) =>
   labwareDef.metadata.displayName
@@ -101,4 +105,25 @@ export function splitWellsOnColumn(
       return [...acc.slice(0, -1), [...lastColumn[0], curr]]
     }
   }, [])
+}
+
+// NOTE: this is used in PD for converting "offset from top" to "mm from bottom".
+// Assumes all wells have same offset because multi-offset not yet supported.
+// TODO: Ian 2019-07-13 return {[string: well]: offset} to support multi-offset
+export const getWellsDepth = (
+  labwareDef: LabwareDefinition2,
+  wells: Array<string>
+): number => {
+  const offsets = wells.map(well => labwareDef.wells[well].depth)
+  if (uniq(offsets).length !== 1) {
+    console.warn(
+      `expected wells ${JSON.stringify(
+        wells
+      )} to all have same offset, but they were different. Labware def is ${getLabwareDefURI(
+        labwareDef
+      )}`
+    )
+  }
+
+  return offsets[0]
 }
